requirements.txt
tensorflow==2.10.0
streamlit==1.22.0
numpy==1.23.5
pillow
